% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyploid_segregation_utils.R
\name{get_dose_segregation}
\alias{get_dose_segregation}
\title{Compute possible gamete dosages from a polyploid genotype}
\usage{
get_dose_segregation(genotype)
}
\arguments{
\item{genotype}{Character. Even-length string of uppercase (reference) and lowercase (alternate) alleles.}
}
\value{
Integer vector of gamete allele counts.
}
\description{
Given a string representation of a polyploid genotype (e.g. “AAaa” for tetraploid),
returns all possible counts of the reference allele (uppercase) in the gametes.
}
\examples{
get_dose_segregation("AAaa")        # tetraploid simplex
get_dose_segregation("AAAaaa")      # hexaploid duplex
}
