% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_data_frames.R
\name{pad_homologs}
\alias{pad_homologs}
\title{Pad Data Frames with Missing Homolog Columns}
\usage{
pad_homologs(df_list, fixed_cols = 5, homolog_prefix = "Homolog_")
}
\arguments{
\item{df_list}{A list of data frames. Each data frame must have at least 4 columns.}

\item{fixed_cols}{The number of fixed columns at the beginning of each data frame.
Defaults to 4.}

\item{homolog_prefix}{The prefix for the homolog columns. Defaults to "homolog_".}
}
\value{
A list of data frames, each padded with NA columns so that they all have the same number of columns.
}
\description{
This function takes a list of data frames that share 4 fixed columns
(assumed to be the first 4 columns) and additional homolog columns
(named "homolog_1", "homolog_2", ...). It pads each data frame with
extra homolog columns filled with NA so that every data frame ends up
having the same total number of columns.
}
