% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pedigree.R
\name{parse_pedigree}
\alias{parse_pedigree}
\title{Parse Pedigree Matrix}
\usage{
parse_pedigree(pedigree)
}
\arguments{
\item{pedigree}{Matrix: Pedigree matrix with three columns: Parent 1, Parent 2, Offspring.}
}
\value{
A list containing:
\describe{
\item{founders}{The row indices of founders in the pedigree.}
\item{crosses}{The row indices of non-founders (crosses) in the pedigree.}
}
}
\description{
This function parses a pedigree matrix and identifies founders and non-founders
(crosses) in the dataset.
}
