% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{plot_difference_matrix}
\alias{plot_difference_matrix}
\title{Plot Difference Matrix Between Two Matrices}
\usage{
plot_difference_matrix(matrix1, matrix2)
}
\arguments{
\item{matrix1}{A reference matrix.}

\item{matrix2}{A matrix to be compared with \code{matrix1}.}
}
\value{
A ggplot object displaying the difference matrix.
}
\description{
This function visualizes differences between two matrices by creating a difference matrix plot.
The plot shows concordant and discordant elements after aligning the columns of \code{matrix2}
with the best matching permutation found by \code{compare_matrices}.
}
