% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMpoly-package.R
\docType{package}
\name{SIMpoly-package}
\alias{SIMpoly}
\alias{SIMpoly-package}
\title{SIMpoly: Simulation of genetic marker data of individuals in pedigreed populations}
\description{
This package simulates genetic marker data of individuals in pedigreed populations. It enables users to model multigenerational, fixed-ploidy populations and analyze allele distributions.
}
\author{
\strong{Maintainer}: Marcelo Mollinari \email{mmollin@ncsu.edu} (\href{https://orcid.org/0000-0002-7001-8498}{ORCID})

}
