% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_offspring.R
\name{simulate_offspring}
\alias{simulate_offspring}
\title{Simulate Offspring from Crosses in Pedigree}
\usage{
simulate_offspring(pedigree, founders, n.mrk, map.length)
}
\arguments{
\item{pedigree}{Matrix: Pedigree matrix with three columns: Parent 1, Parent 2, Offspring.}

\item{founders}{List: Homology groups of founders.}

\item{n.mrk}{Integer: Number of markers.}

\item{map.length}{Numeric: Genetic map length.}
}
\value{
A list of homology groups for each offspring.
}
\description{
This function simulates offspring based on the cross information in a pedigree matrix,
using the homology groups of founders or previous generations.
}
